Rename join function to mjoin & implement new join function
